Create an importer hierarchy to support more types of importers.
