+++ simple/stats/config.py
@@ -12,11 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from stats.data import ImportType
 from stats.data import Provenance
 from stats.data import Source
 from stats.data import StatVar
 
 _INPUT_FILES_FIELD = "inputFiles"
+_IMPORT_TYPE_FIELD = "importType"
 _ENTITY_TYPE_FIELD = "entityType"
 _IGNORE_COLUMNS_FIELD = "ignoreColumns"
 _VARIABLES_FIELD = "variables"
@@ -45,6 +47,15 @@ class Config:
     self.provenance_sources: dict[str, Source] = {}
     self._parse_provenances_and_sources()
 
+  def import_type(self, input_file_name: str) -> ImportType:
+    import_type_str = self._input_file(input_file_name).get(_IMPORT_TYPE_FIELD)
+    if not import_type_str:
+      return ImportType.OBSERVATIONS
+    if import_type_str not in iter(ImportType):
+      raise ValueError(
+          f"Unsupported import type: {import_type_str} ({input_file_name})")
+    return ImportType(import_type_str)
+
   def variable(self, variable_name: str) -> StatVar:
     var_cfg = self.data.get(_VARIABLES_FIELD, {}).get(variable_name, {})
     return StatVar(
@@ -58,7 +69,7 @@ class Config:
   def entity_type(self, input_file_name: str) -> str:
     return self._input_file(input_file_name).get(_ENTITY_TYPE_FIELD, "")
 
-  def ignore_columns(self, input_file_name: str) -> str:
+  def ignore_columns(self, input_file_name: str) -> list[str]:
     return self._input_file(input_file_name).get(_IGNORE_COLUMNS_FIELD, [])
 
   def provenance_name(self, input_file_name: str) -> str:
